Bug: After pausing the timer the timer and progress bar starting from its initial position.
Cause: When timer started initial duration is set same as the duration, but when it is started after pausing, the partial duration is again set for the initial duration, that will make the calculation wrong. since this difference (initial - duration) is used for finding the progress bar length and timer left value (how much it need to move from original starting position) 

Fix: Fixed by calling correct initial duration value based on the completed blocks.